integrating third party libs into browserify. slight project script restructure
