--- js/jquery/.gitignore
@@ -1,3 +1,0 @@
-build
-jquery-migrate.js
-jquery-migrate.min.js
