--- js/jquery/README.md
@@ -1,4 +1,0 @@
-jQuery Component
-================
-
-Shim repository for jQuery.
