--- js/jquery/bower.json
@@ -1,13 +1,0 @@
-{
-  "name": "jquery",
-  "version": "2.0.0",
-  "description": "jQuery component",
-  "keywords": [
-    "jquery",
-    "component"
-  ],
-  "scripts": [
-    "jquery.js"
-  ],
-  "license": "MIT"
-}
