--- js/jquery/component.json
@@ -1,21 +1,0 @@
-{
-  "name": "jquery",
-  "version": "2.0.0",
-  "description": "jQuery component",
-  "keywords": [
-    "jquery",
-    "component"
-  ],
-  "scripts": [
-    "jquery.js"
-  ],
-  "license": "MIT",
-  "gitHead": "46f8412bd1bb9b1b30b5b0eb88560e2d4196509c",
-  "readme": "jQuery Component\n================\n\nShim repository for jQuery.\n",
-  "readmeFilename": "README.md",
-  "_id": "jquery@2.0.0",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/components/jquery.git"
-  }
-}
