--- js/jquery/composer.json
@@ -1,30 +1,0 @@
-{
-    "name": "components/jquery",
-    "description": "jQuery JavaScript Library",
-    "type": "component",
-    "homepage": "http://jquery.com",
-    "license": "MIT",
-    "support": {
-        "irc": "irc://irc.freenode.org/jquery",
-        "issues": "http://bugs.jquery.com",
-        "forum": "http://forum.jquery.com",
-        "wiki": "http://docs.jquery.com/",
-        "source": "https://github.com/jquery/jquery"
-    },
-    "authors": [
-        {
-            "name": "John Resig",
-            "email": "jeresig@gmail.com"
-        }
-    ],
-    "require": {
-        "robloach/component-installer": "*"
-    },
-    "extra": {
-        "component": {
-            "scripts": [
-                "jquery.js"
-            ]
-        }
-    }
-}
