--- js/jquery/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "components-jquery",
-  "version": "2.0.0",
-  "description": "jQuery component",
-  "keywords": ["jquery"],
-  "main": "./jquery.js"
-}
