+++ src/analytics.js
@@ -1,21 +1,23 @@
-var init = require('./../init/init.js');
+var app = require('./app.js');
+
+var $ = require('jquery');
 
 (function(e,t){var r=e.amplitude||{};var n=t.createElement("script");n.type="text/javascript";
   n.async=true;n.src="https://d24n15hnbwhuhn.cloudfront.net/libs/amplitude-2.0.4-min.js";
   var s=t.getElementsByTagName("script")[0];s.parentNode.insertBefore(n,s);r._q=[];
   function a(e){r[e]=function(){r._q.push([e].concat(Array.prototype.slice.call(arguments,0)))
-  }}var i=["init","logEvent","setUserId","setUserProperties","setOptOut","setVersionName","setDomain","setDeviceId","setGlobalUserProperties"];
+  }}var i=["app","logEvent","setUserId","setUserProperties","setOptOut","setVersionName","setDomain","setDeviceId","setGlobalUserProperties"];
   for(var o=0;o<i.length;o++){a(i[o])}e.amplitude=r}
 )(window,document);
 
-amplitude.init("{{amplitude_key}}");
+amplitude.app("{{amplitude_key}}");
 
 var setUser = function(){
   var instagramUrl = "https://api.instagram.com/v1/users/self/";
   $.ajax({
     url: instagramUrl,
     data: {
-      access_token: init.authCode
+      access_token: app.authCode
     },
     success: function(response){
       amplitude.setUserId(response.data.username);
@@ -25,8 +27,9 @@ var setUser = function(){
 };
 
 var eventProperties = {
-  appVersion: init.app.version
+  appVersion: app.app.version
 };
+
 amplitude.logEvent("opened new tab", eventProperties);
 
 // fixes amplitude unsent errors
