+++ src/auth/finished.html
@@ -12,8 +12,6 @@
 			Whenever you open a new tab, you'll see photos from your Instagram feed! <br>
 	  </div>
 	</div>
-	<script charset="utf-8" src="/js/moment.js"></script>
-	<script charset="utf-8" src="/js/jquery/jquery.min.js"></script>
 	<script charset="utf-8" src="/dist/build.js"></script>
 </body>
 </html>
