+++ src/index.js
@@ -1,0 +1,4 @@
+require('./auth/finished.js');
+require('./override/tab.js');
+require('./options/options.js');
+require('./analytics.js');
