+++ src/options/index.html
@@ -69,9 +69,6 @@
 		InstaTab is not related to or affiliated with Instagram or Facebook.<br>
 		Instagram users own their content, this extension does not claim any ownership of any content.<br>
 	</div>
-	<script charset="utf-8" src="/js/moment.js"></script>
-	<script charset="utf-8" src="/js/jquery/jquery.min.js"></script>
-	<script charset="utf-8" src="/js/D.min.js"></script>
 	<script charset="utf-8" src="/dist/build.js"></script>]
 </body>
 </html>
