+++ src/override/override.html
@@ -17,9 +17,6 @@
 		<div class="first-row"></div>
 		<div class="second-row"></div>
 	</div>
-	<script charset="utf-8" src="/js/moment.js"></script>
-	<script charset="utf-8" src="/js/jquery/jquery.min.js"></script>
-	<script charset="utf-8" src="/js/D.min.js"></script>
 	<script charset="utf-8" src="/dist/build.js"></script>
 </body>
 </html>
